moved tools install
