+++ src-tauri/src/gui/commands/mod.rs
@@ -2,3 +2,4 @@ pub mod utils_commands;
 pub mod prequisites;
 pub mod installation;
 pub mod settings;
+pub mod idf_tools;
